backend/requirements.txt
Django==3.2.20
djangorestframework==3.14.0
djongo==1.3.6
django-cors-headers==4.1.0
pymongo==3.12.3
python-dotenv==1.0.0
sqlparse==0.2.4
djangorestframework-simplejwt
requests==2.31.0
aiohttp==3.9.3
beautifulsoup4==4.12.3
